Refactor various things

+++ glgrib_field_scalar.h
@@ -10,7 +10,7 @@ public:
   glgrib_field_scalar & operator= (const glgrib_field_scalar &);
   glgrib_field_scalar () { }
   glgrib_field_scalar (const glgrib_field_scalar &);
-  virtual void init (const glgrib_options_field &, int = 0);
+  virtual void init (const glgrib_options_field &, float = 0);
   virtual void render (const glgrib_view &, const glgrib_options_light &) const;
   virtual ~glgrib_field_scalar ();
   void setupVertexAttributes ();
