+++ glgrib_field_vector.h
@@ -11,7 +11,7 @@ public:
   glgrib_field_vector & operator= (const glgrib_field_vector &);
   glgrib_field_vector () { }
   glgrib_field_vector (const glgrib_field_vector &);
-  virtual void init (const glgrib_options_field &, int = 0);
+  virtual void init (const glgrib_options_field &, float = 0);
   virtual void render (const glgrib_view &, const glgrib_options_light &) const;
   virtual ~glgrib_field_vector ();
   void setupVertexAttributes ();
